requirements.txt
streamlit==1.33.0
tensorflow==2.17.0
keras==3.4.1
pillow==10.3.0
 